plain text diff in email
Using pickle as storage, plain text is nice though, but hard to write. Now, basically we have a dict, which keys are urls holding the latest hash and the corresponding compresed website. When diffing, we uncompress and show the difference and compress the new website and hash.

Hint: it could consume some memory to store every diff locally in one file, but it's zlib'd, which give us only ~ 1/3 of the original file.