Add check_interval function
The last-check timestamps are stored in cache.yml in the diff and hash
cache directory - if the file does not exist, it will be created.
If check_interval was missing in conf.yml it's value defaults to 5 minutes.
(Remember to execute ´wcc´ that recently to allow wcc satisfying your intervals.)