Extend HTTP error handling of redirects
On temporary redirect wcc silently (message on WARN level) follows exactly
one redirection an takes that result as the final one.

On permanent redirect wcc issues an ERROR message containing the information
which site (if any) is considered the new one and challanges the user to update
his conf.yml entry accordingly.